Extract backpack code to a device
