+++ src/lib/Backpack/devBackpack.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "device.h"
+
+extern device_t Backpack_device;
